# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nigerian School Report Card Generator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /schools/{schoolId}/terms:
    get:
      operationId: listTerms
      tags: [terms]
      summary: List academic terms for a school
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of terms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AcademicTerm"
    post:
      operationId: createTerm
      tags: [terms]
      summary: Create an academic term
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcademicTermInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcademicTerm"
  /terms/{id}:
    patch:
      operationId: updateTerm
      tags: [terms]
      summary: Update an academic term
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcademicTermUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcademicTerm"
    delete:
      operationId: deleteTerm
      tags: [terms]
      summary: Delete a term
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── SUBJECTS ───────────────────────────────────────────────────────────────
  /schools/{schoolId}/subjects:
    get:
      operationId: listSubjects
      tags: [subjects]
      summary: List subjects for a school
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
    post:
      operationId: createSubject
      tags: [subjects]
      summary: Create a subject
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
  /subjects/{id}:
    patch:
      operationId: updateSubject
      tags: [subjects]
      summary: Update a subject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    delete:
      operationId: deleteSubject
      tags: [subjects]
      summary: Delete a subject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── SUBJECT SCORES ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    AcademicTerm:
      type: object
      required: [id, schoolId, session, term]
      properties:
        id:         { type: integer }
        schoolId:   { type: integer }
        session:    { type: string }
        term:       { type: integer }
        startDate:  { type: ["string", "null"] }
        endDate:    { type: ["string", "null"] }
        nextTermBegins: { type: ["string", "null"] }
    AcademicTermInput:
      type: object
      required: [session, term]
      properties:
        session:    { type: string }
        term:       { type: integer }
        startDate:  { type: string }
        endDate:    { type: string }
        nextTermBegins: { type: string }
    AcademicTermUpdate:
      type: object
      properties:
        session:    { type: string }
        term:       { type: integer }
        startDate:  { type: string }
        endDate:    { type: string }
        nextTermBegins: { type: string }

    # ── Subjects ──
    Subject:
      type: object
      required: [id, schoolId, name]
      properties:
        id:       { type: integer }
        schoolId: { type: integer }
        name:     { type: string }
        code:     { type: ["string", "null"] }
    SubjectInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        code: { type: string }
    SubjectUpdate:
      type: object
      properties:
        name: { type: string }
        code: { type: string }

    # ── Subject Scores ──
...